Compose webpack config, restructure js
- Composed webpack config to a dev config and production config.

- Modified .gitignore to ignore ./dist

- Restructured index.js so that it's just doing imports

- Added Upload.js which has all of the functions required for handling
file uploads
