+++ .gitignore
@@ -2,6 +2,8 @@
 
 /example/*
 
+/dist/
+
 # npm modules
 
 /node_modules/*
