+++ package.json
@@ -5,8 +5,8 @@
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --mode=development --config config/webpack.config.js",
-    "build": "webpack --env production --config config/webpack.config.js"
+    "start": "webpack-dev-server --mode=development --config config/webpack.dev.js",
+    "build": "webpack --env production --mode=production --config config/webpack.prod.js"
   },
   "repository": {
     "type": "git",
