--- src/stylesheets/main.css
@@ -1,92 +1,0 @@
-main {
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-columns: 1fr 2fr 2fr;
-      grid-template-columns: 1fr 2fr 2fr;
-  -ms-grid-rows: 3fr 2fr 3fr;
-      grid-template-rows: 3fr 2fr 3fr;
-      grid-template-areas: "intro canvas canvas" "intro canvas canvas" "anc anc anc";
-  height: 100vh;
-}
-
-.intro {
-  -ms-grid-row: 1;
-  -ms-grid-row-span: 2;
-  -ms-grid-column: 1;
-  grid-area: intro;
-}
-
-.js-danger-popup {
-  padding: 10px;
-  color: white;
-  background: red;
-  font-weight: bold;
-}
-
-.drop__target {
-  width: 200px;
-  height: 200px;
-  border: 3px dashed grey;
-}
-
-.upload__row {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-flow: column nowrap;
-          flex-flow: column nowrap;
-}
-
-.upload__label {
-  display: block;
-}
-
-.button--upload {
-  display: block;
-}
-
-.strip {
-  list-style: none;
-}
-
-.canvas-wrapper {
-  -ms-grid-row: 1;
-  -ms-grid-row-span: 2;
-  -ms-grid-column: 2;
-  -ms-grid-column-span: 2;
-  grid-area: canvas;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-}
-
-.ancillary {
-  -ms-grid-row: 3;
-  -ms-grid-column: 1;
-  -ms-grid-column-span: 3;
-  grid-area: anc;
-}
-
-.strip {
-  grid-area: strip;
-  list-style: none;
-  padding: 0;
-}
-
-.strip__item {
-  display: inline-block;
-}
-
-.strip__img {
-  max-width: 20vh;
-  height: auto;
-}
-/*# sourceMappingURL=main.css.map */
